feat: sort HDF_PATH glob results by modification time
The recipe globs for XSA files specified in HDF_PATH when that contains
an asterisk (*). Without this change, the results are sorted
alphabetically and the first item in the sorted list is chosen. This
will likely be the lowest build number, thus the oldest build.

Usually a developer wants to build with the latest version of the XSA
file. This could be the highest build number, but it doesn't have to be.
Sorting by the modification time seems like a safer approach.